rip out of WD intro extra module
* SpecificationRepository is now derivable, special class for WD
* Profile is not in charge of SpecificationRepository creation
