+++ key.core.wd/src/main/java/de/uka/ilkd/key/wd/LoopWellDefinedness.java
