+++ key.core.wd/src/main/java/de/uka/ilkd/key/wd/StatementWellDefinedness.java
@@ -12,9 +12,9 @@ import de.uka.ilkd.key.logic.JTerm;
 import de.uka.ilkd.key.logic.TermBuilder;
 import de.uka.ilkd.key.logic.TermServices;
 import de.uka.ilkd.key.logic.op.*;
-import de.uka.ilkd.key.proof.init.WellDefinednessPO.Variables;
 import de.uka.ilkd.key.speclang.ContractFactory;
 import de.uka.ilkd.key.speclang.SpecificationElement;
+import de.uka.ilkd.key.wd.po.WellDefinednessPO.Variables;
 
 import org.key_project.logic.op.Function;
 import org.key_project.prover.sequent.SequentFormula;
