+++ key.core.wd/src/main/java/de/uka/ilkd/key/wd/WellDefinednessCheck.java
@@ -21,14 +21,14 @@ import de.uka.ilkd.key.proof.OpReplacer;
 import de.uka.ilkd.key.proof.init.ContractPO;
 import de.uka.ilkd.key.proof.init.InitConfig;
 import de.uka.ilkd.key.proof.init.ProofOblInput;
-import de.uka.ilkd.key.proof.init.WellDefinednessPO;
-import de.uka.ilkd.key.proof.init.WellDefinednessPO.Variables;
 import de.uka.ilkd.key.rule.RewriteTaclet;
 import de.uka.ilkd.key.rule.tacletbuilder.RewriteTacletBuilder;
 import de.uka.ilkd.key.settings.ProofSettings;
 import de.uka.ilkd.key.speclang.Contract;
 import de.uka.ilkd.key.speclang.jml.JMLInfoExtractor;
 import de.uka.ilkd.key.util.MiscTools;
+import de.uka.ilkd.key.wd.po.WellDefinednessPO;
+import de.uka.ilkd.key.wd.po.WellDefinednessPO.Variables;
 
 import org.key_project.logic.Name;
 import org.key_project.logic.op.Function;
