+++ key.core.wd/src/main/resources/META-INF/services/de.uka.ilkd.key.proof.init.DefaultProfileResolver
@@ -1,0 +1,1 @@
+de.uka.ilkd.key.wd.WdProfileResolver
