+++ key.core/src/main/java/de/uka/ilkd/key/java/Services.java
@@ -103,7 +103,7 @@ public class Services implements TermServices, LogicServices, ProofServices {
         this.caches = new ServiceCaches();
         this.termBuilder = new TermBuilder(new TermFactory(caches.getTermFactoryCache()), this);
         this.termBuilderWithoutCache = new TermBuilder(new TermFactory(), this);
-        this.specRepos = new SpecificationRepository(this);
+        this.specRepos = profile.createSpecificationRepository(this);
         cee = new ConstantExpressionEvaluator(this);
         typeconverter = new TypeConverter(this);
         javainfo = new JavaInfo(
@@ -122,7 +122,7 @@ public class Services implements TermServices, LogicServices, ProofServices {
         this.caches = caches;
         this.termBuilder = new TermBuilder(new TermFactory(caches.getTermFactoryCache()), this);
         this.termBuilderWithoutCache = new TermBuilder(new TermFactory(), this);
-        this.specRepos = new SpecificationRepository(this);
+        this.specRepos = profile.createSpecificationRepository(this);
         cee = new ConstantExpressionEvaluator(this);
         typeconverter = new TypeConverter(this);
         javainfo = new JavaInfo(new KeYProgModelInfo(this, crsc, rec2key, typeconverter), this);
