+++ key.core/src/main/java/de/uka/ilkd/key/proof/init/AbstractOperationPO.java
@@ -456,7 +456,7 @@ public abstract class AbstractOperationPO extends AbstractPO {
         collectClassAxioms(getCalleeKeYJavaType(), proofConfig);
 
         // for JML annotation statements
-        generateWdTaclets(proofConfig);
+        generateDynamicTaclets(proofConfig);
     }
 
     /**
