+++ key.core/src/main/java/de/uka/ilkd/key/proof/init/FunctionalBlockContractPO.java
@@ -322,7 +322,7 @@ public class FunctionalBlockContractPO extends AbstractPO implements ContractPO 
         assignPOTerms(validity);
 
         collectClassAxioms(getCalleeKeYJavaType(), proofConfig);
-        generateWdTaclets(proofConfig);
+        generateDynamicTaclets(proofConfig);
     }
 
     /**
