+++ key.core/src/main/java/de/uka/ilkd/key/proof/init/FunctionalLoopContractPO.java
@@ -177,7 +177,7 @@ public class FunctionalLoopContractPO extends AbstractPO implements ContractPO {
 
         assignPOTerms(validity);
         collectClassAxioms(getCalleeKeYJavaType(), proofConfig);
-        generateWdTaclets(proofConfig);
+        generateDynamicTaclets(proofConfig);
     }
 
     /**
