+++ key.core/src/main/java/de/uka/ilkd/key/proof/mgt/SpecificationRepository.java
@@ -34,10 +34,6 @@ import de.uka.ilkd.key.speclang.jml.JMLInfoExtractor;
 import de.uka.ilkd.key.speclang.jml.translation.ProgramVariableCollection;
 import de.uka.ilkd.key.speclang.translation.SLTranslationException;
 import de.uka.ilkd.key.util.MiscTools;
-import de.uka.ilkd.key.wd.ClassWellDefinedness;
-import de.uka.ilkd.key.wd.MethodWellDefinedness;
-import de.uka.ilkd.key.wd.StatementWellDefinedness;
-import de.uka.ilkd.key.wd.WellDefinednessCheck;
 
 import org.key_project.logic.Name;
 import org.key_project.logic.sort.Sort;
@@ -45,11 +41,7 @@ import org.key_project.prover.rules.ApplicationRestriction;
 import org.key_project.prover.rules.RuleSet;
 import org.key_project.prover.sequent.Sequent;
 import org.key_project.prover.sequent.SequentFormula;
-import org.key_project.util.collection.DefaultImmutableSet;
-import org.key_project.util.collection.ImmutableList;
-import org.key_project.util.collection.ImmutableSLList;
-import org.key_project.util.collection.ImmutableSet;
-import org.key_project.util.collection.Pair;
+import org.key_project.util.collection.*;
 
 import org.jspecify.annotations.Nullable;
 import org.slf4j.Logger;
@@ -60,7 +52,7 @@ import org.slf4j.LoggerFactory;
  * invariants. Provides methods for adding such elements to the repository, and for retrieving them
  * afterwards.
  */
-public final class SpecificationRepository {
+public class SpecificationRepository {
     private static final Logger LOGGER = LoggerFactory.getLogger(SpecificationRepository.class);
 
     public static final String CONTRACT_COMBINATION_MARKER = "#";
@@ -74,17 +66,15 @@ public final class SpecificationRepository {
 
     private final ContractFactory cf;
 
-    private final Map<Pair<KeYJavaType, IObserverFunction>, ImmutableSet<Contract>> contracts =
+    protected final Map<Pair<KeYJavaType, IObserverFunction>, ImmutableSet<Contract>> contracts =
         new LinkedHashMap<>();
-    private final Map<Pair<KeYJavaType, IProgramMethod>, ImmutableSet<FunctionalOperationContract>> operationContracts =
+    protected final Map<Pair<KeYJavaType, IProgramMethod>, ImmutableSet<FunctionalOperationContract>> operationContracts =
         new LinkedHashMap<>();
-    private final Map<Pair<KeYJavaType, IObserverFunction>, ImmutableSet<WellDefinednessCheck>> wdChecks =
-        new LinkedHashMap<>();
-    private final Map<String, Contract> contractsByName = new LinkedHashMap<>();
-    private final Map<KeYJavaType, ImmutableSet<IObserverFunction>> contractTargets =
+    protected final Map<String, Contract> contractsByName = new LinkedHashMap<>();
+    protected final Map<KeYJavaType, ImmutableSet<IObserverFunction>> contractTargets =
         new LinkedHashMap<>();
     private final Map<KeYJavaType, ImmutableSet<ClassInvariant>> invs = new LinkedHashMap<>();
-    private final Map<KeYJavaType, ImmutableSet<ClassAxiom>> axioms = new LinkedHashMap<>();
+    protected final Map<KeYJavaType, ImmutableSet<ClassAxiom>> axioms = new LinkedHashMap<>();
     private final Map<KeYJavaType, ImmutableSet<InitiallyClause>> initiallyClauses =
         new LinkedHashMap<>();
     private final Map<ProofOblInput, ImmutableSet<Proof>> proofs = new LinkedHashMap<>();
@@ -124,8 +114,8 @@ public final class SpecificationRepository {
     private final Map<KeYJavaType, ImmutableSet<ClassAxiom>> allClassAxiomsCache =
         new LinkedHashMap<>();
 
-    private final Services services;
-    private final TermBuilder tb;
+    protected final Services services;
+    protected final TermBuilder tb;
 
     private final Map<String, Integer> contractCounters =
         new de.uka.ilkd.key.util.LinkedHashMap<>();
@@ -223,7 +213,7 @@ public final class SpecificationRepository {
         }
     }
 
-    private IObserverFunction getCanonicalFormForKJT(IObserverFunction obs, KeYJavaType kjt) {
+    protected IObserverFunction getCanonicalFormForKJT(IObserverFunction obs, KeYJavaType kjt) {
         assert obs != null;
         assert kjt != null;
         if (!(obs instanceof IProgramMethod pm) || obs.getContainerType().equals(kjt)) {
@@ -375,7 +365,7 @@ public final class SpecificationRepository {
         return contract;
     }
 
-    private void registerContract(Contract contract) {
+    protected void registerContract(Contract contract) {
         final Pair<KeYJavaType, IObserverFunction> target =
             new Pair<>(contract.getKJT(), contract.getTarget());
         registerContract(contract, target);
@@ -388,12 +378,9 @@ public final class SpecificationRepository {
         }
     }
 
-    private void registerContract(Contract contract,
+    protected void registerContract(Contract contract,
             Pair<KeYJavaType, IObserverFunction> targetPair) {
         LOGGER.trace("Contract registered {}", contract);
-        if (!WellDefinednessCheck.isOn() && contract instanceof WellDefinednessCheck) {
-            return;
-        }
         final KeYJavaType targetKJT = targetPair.first;
         final IObserverFunction targetMethod = targetPair.second;
         contract = contract.setTarget(targetKJT, targetMethod);
@@ -406,62 +393,24 @@ public final class SpecificationRepository {
         assert contract.id() != Contract.INVALID_ID : "Tried to add a contract with an invalid id!";
         contracts.put(targetPair, getContracts(targetKJT, targetMethod).add(contract));
 
-        if (contract instanceof FunctionalOperationContract) {
-            operationContracts.put(new Pair<>(targetKJT, (IProgramMethod) targetMethod),
-                getOperationContracts(targetKJT, (IProgramMethod) targetMethod)
-                        .add((FunctionalOperationContract) contract));
-            // Create new well-definedness check
-            final MethodWellDefinedness mwd =
-                new MethodWellDefinedness((FunctionalOperationContract) contract, services);
-            registerContract(mwd);
-        } else if (contract instanceof DependencyContract && contract.getOrigVars().atPres.isEmpty()
-                && targetMethod.getContainerType()
-                        .equals(services.getJavaInfo().getJavaLangObject())) {
-            // Create or extend a well-definedness check for a class invariant
-            final JTerm deps =
-                contract.getAccessible(services.getTypeConverter().getHeapLDT().getHeap());
-            final JTerm mby = contract.getMby();
-            final String invName = "JML model class invariant in " + targetKJT.getName();
-            final ClassInvariant inv = new ClassInvariantImpl(invName, invName, targetKJT,
-                contract.getVisibility(), tb.tt(), contract.getOrigVars().self);
-            ClassWellDefinedness cwd =
-                new ClassWellDefinedness(inv, targetMethod, deps, mby, services);
-            final ImmutableSet<ClassWellDefinedness> cwds = getWdClassChecks(targetKJT);
-            if (!cwds.isEmpty()) {
-                assert cwds.size() == 1;
-                final ClassWellDefinedness oldCwd = cwds.iterator().next();
-                unregisterContract(oldCwd);
-                oldCwd.addInv(cwd.getInvariant().getInv(oldCwd.getOrigVars().self, services));
-                cwd = oldCwd.combine(cwd, services);
-            }
-            registerContract(cwd);
-        } else if (contract instanceof DependencyContract
-                && contract.getOrigVars().atPres.isEmpty()) {
-            // Create or extend a well-definedness check for a model field
-            MethodWellDefinedness mwd =
-                new MethodWellDefinedness((DependencyContract) contract, services);
-            final ImmutableSet<MethodWellDefinedness> mwds =
-                getWdMethodChecks(targetKJT, targetMethod);
-            if (!mwds.isEmpty()) {
-                assert mwds.size() == 1;
-                final MethodWellDefinedness oldMwd = mwds.iterator().next();
-                unregisterContract(oldMwd);
-                mwd = mwd.combine(oldMwd, services);
-            }
-            registerContract(mwd);
-        } else if (contract instanceof WellDefinednessCheck) {
-            registerWdCheck((WellDefinednessCheck) contract);
-        }
         contractsByName.put(contract.getName(), contract);
         final ImmutableSet<IObserverFunction> oldTargets = getContractTargets(targetKJT);
         final ImmutableSet<IObserverFunction> newTargets = oldTargets.add(targetMethod);
         contractTargets.put(targetKJT, newTargets);
+
+        if (contract instanceof FunctionalOperationContract operationContract) {
+            operationContracts.put(new Pair<>(targetKJT, (IProgramMethod) targetMethod),
+                getOperationContracts(targetKJT, (IProgramMethod) targetMethod)
+                        .add(operationContract));
+        } else {
+            LOGGER.warn("Unknown contract handed in: {}", contract);
+        }
     }
 
     /**
      * Removes the contract from the repository, but keeps its target.
      */
-    private void unregisterContract(Contract contract) {
+    protected void unregisterContract(Contract contract) {
         final KeYJavaType kjt = contract.getKJT();
         final Pair<KeYJavaType, IObserverFunction> tp = new Pair<>(kjt, contract.getTarget());
         contracts.put(tp, contracts.get(tp).remove(contract));
@@ -470,19 +419,6 @@ public final class SpecificationRepository {
                 new Pair<>(tp.first, (IProgramMethod) tp.second);
             operationContracts.put(tp2,
                 operationContracts.get(tp2).remove((FunctionalOperationContract) contract));
-            if (!getWdChecks(contract.getKJT(), contract.getTarget()).isEmpty()) {
-                ImmutableSet<WellDefinednessCheck> wdcs =
-                    getWdChecks(contract.getKJT(), contract.getTarget());
-                for (WellDefinednessCheck wdc : wdcs) {
-                    if (wdc instanceof MethodWellDefinedness
-                            && ((MethodWellDefinedness) wdc).getMethodContract().equals(contract)) {
-                        unregisterWdCheck(wdc);
-                    }
-                }
-            }
-        }
-        if (contract instanceof WellDefinednessCheck) {
-            unregisterWdCheck((WellDefinednessCheck) contract);
         }
         contractsByName.remove(contract.getName());
     }
@@ -513,8 +449,8 @@ public final class SpecificationRepository {
                 if (oldFuncContracts.isEmpty()) {
                     final FunctionalOperationContract iniContr = cf.func(pm, inv);
                     addContractNoInheritance(iniContr);
-                    assert getContracts(kjt, pm).size() == (WellDefinednessCheck.isOn() ? 2 : 1)
-                            + oldContracts.size();
+                    // assert getContracts(kjt, pm).size() == (WellDefinednessCheck.isOn() ? 2 : 1)
+                    // + oldContracts.size();
                 } else {
                     for (FunctionalOperationContract c : oldFuncContracts) {
                         unregisterContract(c);
@@ -530,132 +466,6 @@ public final class SpecificationRepository {
     // Internal interface for well-definedness checks
 
     /**
-     * Remove well-definedness checks from a given set of contracts
-     *
-     * @param contracts A set of contracts
-     * @return contracts without well-definedness checks
-     */
-    private static ImmutableSet<Contract> removeWdChecks(ImmutableSet<Contract> contracts) {
-        ImmutableList<Contract> result = ImmutableSLList.nil();
-        if (contracts == null) {
-            return contracts;
-        }
-        for (Contract c : contracts) {
-            if (!(c instanceof WellDefinednessCheck)) {
-                result = result.prepend(c);
-            }
-        }
-        return DefaultImmutableSet.fromImmutableList(result);
-    }
-
-    /**
-     * Registers a well-definedness check. It does not take care of its visibility in the proof
-     * management dialog (this is done in {@link #registerContract(Contract, Pair)}).
-     *
-     * @param check The well-definedness check to be registered
-     */
-    private void registerWdCheck(WellDefinednessCheck check) {
-        ImmutableSet<WellDefinednessCheck> checks =
-            getWdChecks(check.getKJT(), check.getTarget()).add(check);
-        wdChecks.put(new Pair<>(check.getKJT(), check.getTarget()), checks);
-    }
-
-    /**
-     * Unregisters a well-definedness check. It does not take care of its visibility in the proof
-     * management dialog.
-     *
-     * @param check The well-definedness check to be unregistered
-     */
-    private void unregisterWdCheck(WellDefinednessCheck check) {
-        wdChecks.put(new Pair<>(check.getKJT(), check.getTarget()),
-            getWdChecks(check.getKJT(), check.getTarget()).remove(check));
-    }
-
-    /**
-     * Returns all registered (atomic) well-definedness checks for the passed kjt.
-     */
-    private ImmutableSet<WellDefinednessCheck> getWdChecks(KeYJavaType kjt) {
-        assert kjt != null;
-        ImmutableSet<WellDefinednessCheck> result = DefaultImmutableSet.nil();
-        for (WellDefinednessCheck ch : getAllWdChecks()) {
-            if (ch.getKJT().equals(kjt)) {
-                result = result.add(ch);
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Returns all registered (atomic) well-definedness checks for the passed target and kjt.
-     */
-    private ImmutableSet<WellDefinednessCheck> getWdChecks(KeYJavaType kjt,
-            IObserverFunction target) {
-        assert kjt != null;
-        assert target != null;
-        target = getCanonicalFormForKJT(target, kjt);
-        final Pair<KeYJavaType, IObserverFunction> pair = new Pair<>(kjt, target);
-        final ImmutableSet<WellDefinednessCheck> result = wdChecks.get(pair);
-        return result == null ? DefaultImmutableSet.nil() : result;
-    }
-
-    /**
-     * Returns all registered well-definedness checks for method contracts.
-     */
-    private ImmutableSet<MethodWellDefinedness> getAllWdMethodChecks() {
-        ImmutableSet<MethodWellDefinedness> result = DefaultImmutableSet.nil();
-        for (var s : getAllWdChecks()) {
-            if (s instanceof MethodWellDefinedness) {
-                result = result.add((MethodWellDefinedness) s);
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Returns all registered (atomic) well-definedness method checks for the passed kjt.
-     */
-    private ImmutableSet<MethodWellDefinedness> getWdMethodChecks(KeYJavaType kjt) {
-        assert kjt != null;
-        ImmutableSet<MethodWellDefinedness> result = DefaultImmutableSet.nil();
-        for (MethodWellDefinedness ch : getAllWdMethodChecks()) {
-            if (ch.getKJT().equals(kjt)) {
-                result = result.add(ch);
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Returns all registered (atomic) well-definedness method checks for the passed target and kjt.
-     */
-    private ImmutableSet<MethodWellDefinedness> getWdMethodChecks(KeYJavaType kjt,
-            IObserverFunction target) {
-        assert kjt != null;
-        assert target != null;
-        ImmutableSet<MethodWellDefinedness> result = DefaultImmutableSet.nil();
-        for (MethodWellDefinedness ch : getAllWdMethodChecks()) {
-            if (ch.getKJT().equals(kjt) && ch.getTarget().equals(target)) {
-                result = result.add(ch);
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Returns all registered (atomic) well-definedness class checks for the passed kjt.
-     */
-    private ImmutableSet<ClassWellDefinedness> getWdClassChecks(KeYJavaType kjt) {
-        ImmutableSet<WellDefinednessCheck> checks = getWdChecks(kjt);
-        ImmutableSet<ClassWellDefinedness> invs = DefaultImmutableSet.nil();
-        for (WellDefinednessCheck check : checks) {
-            if (check instanceof ClassWellDefinedness) {
-                invs = invs.add((ClassWellDefinedness) check);
-            }
-        }
-        return invs;
-    }
-
-    /**
      * Helper for {@link #map(UnaryOperator, Services)}.
      *
      * @param map a map.
@@ -663,7 +473,7 @@ public final class SpecificationRepository {
      * @param services services.
      */
     @SuppressWarnings("unchecked")
-    private <K, V extends SpecificationElement> void mapValueSets(Map<K, ImmutableSet<V>> map,
+    protected <K, V extends SpecificationElement> void mapValueSets(Map<K, ImmutableSet<V>> map,
             UnaryOperator<JTerm> op, Services services) {
         for (Entry<K, ImmutableSet<V>> entry : map.entrySet()) {
             final K key = entry.getKey();
@@ -724,7 +534,6 @@ public final class SpecificationRepository {
     public void map(UnaryOperator<JTerm> op, Services services) {
         mapValueSets(contracts, op, services);
         mapValueSets(operationContracts, op, services);
-        mapValueSets(wdChecks, op, services);
         mapValueSets(invs, op, services);
         mapValueSets(axioms, op, services);
         mapValueSets(initiallyClauses, op, services);
@@ -743,7 +552,7 @@ public final class SpecificationRepository {
         for (ImmutableSet<Contract> s : contracts.values()) {
             result = result.union(s);
         }
-        return WellDefinednessCheck.isOn() ? result : removeWdChecks(result);
+        return result;
     }
 
     /**
@@ -753,8 +562,7 @@ public final class SpecificationRepository {
         target =
             getCanonicalFormForKJT(Objects.requireNonNull(target), Objects.requireNonNull(kjt));
         final Pair<KeYJavaType, IObserverFunction> pair = new Pair<>(kjt, target);
-        final ImmutableSet<Contract> result =
-            WellDefinednessCheck.isOn() ? contracts.get(pair) : removeWdChecks(contracts.get(pair));
+        final ImmutableSet<Contract> result = contracts.get(pair);
         return result == null ? DefaultImmutableSet.nil() : result;
     }
 
@@ -943,17 +751,6 @@ public final class SpecificationRepository {
         final IObserverFunction target = inv.isStatic() ? services.getJavaInfo().getStaticInv(kjt)
                 : services.getJavaInfo().getInv();
         invs.put(kjt, getClassInvariants(kjt).add(inv));
-        final ImmutableSet<ClassWellDefinedness> cwds = getWdClassChecks(kjt);
-        if (cwds.isEmpty()) {
-            registerContract(new ClassWellDefinedness(inv, target, null, null, services));
-        } else {
-            assert cwds.size() == 1;
-            ClassWellDefinedness cwd = cwds.iterator().next();
-            unregisterContract(cwd);
-            cwd = cwd.combine(new ClassWellDefinedness(inv, cwd.getTarget(), null, null, services),
-                services);
-            registerContract(cwd);
-        }
 
         // in any case, create axiom with non-static target
         addClassAxiom(new PartialInvAxiom(inv, false, services));
@@ -966,21 +763,7 @@ public final class SpecificationRepository {
             final ImmutableList<KeYJavaType> subs = services.getJavaInfo().getAllSubtypes(kjt);
             for (KeYJavaType sub : subs) {
                 ClassInvariant subInv = inv.setKJT(sub);
-                final IObserverFunction subTarget =
-                    subInv.isStatic() ? services.getJavaInfo().getStaticInv(sub)
-                            : services.getJavaInfo().getInv();
                 invs.put(sub, getClassInvariants(sub).add(subInv));
-                final ImmutableSet<ClassWellDefinedness> subCwds = getWdClassChecks(sub);
-                if (subCwds.isEmpty()) {
-                    registerContract(
-                        new ClassWellDefinedness(subInv, subTarget, null, null, services));
-                } else {
-                    for (ClassWellDefinedness cwd : subCwds) {
-                        unregisterContract(cwd);
-                        cwd.addInv(subInv.getInv(cwd.getOrigVars().self, services));
-                        registerContract(cwd);
-                    }
-                }
             }
         }
     }
@@ -1804,65 +1587,6 @@ public final class SpecificationRepository {
         return result;
     }
 
-    // Public interface for well-definedness checks
-
-    /**
-     * Represent terms belong to model fields, so the well-definedness check considers both of them
-     * together.
-     *
-     * @param kjt The relevant KeYJavaType
-     */
-    public void addRepresentsTermToWdChecksForModelFields(KeYJavaType kjt) {
-        ImmutableSet<ClassAxiom> axs = axioms.get(kjt);
-        if (axs == null) {
-            return;
-        }
-        ImmutableSet<RepresentsAxiom> reps = DefaultImmutableSet.nil();
-        for (ClassAxiom ax : axs) {
-            if (ax instanceof RepresentsAxiom) {
-                reps = reps.add((RepresentsAxiom) ax);
-            }
-        }
-        final ProgramVariable heap = services.getTypeConverter().getHeapLDT().getHeap();
-        for (RepresentsAxiom rep : reps) {
-            boolean dep = false;
-            for (MethodWellDefinedness ch : getWdMethodChecks(kjt)) {
-                if (ch.modelField() && ch.getTarget().equals(rep.getTarget())) {
-                    dep = true;
-                    unregisterContract(ch);
-                    JTerm represents = rep.getAxiom(heap, ch.getOrigVars().self, services);
-                    WellDefinednessCheck newCh = ch.addRepresents(represents);
-                    registerContract(newCh);
-                }
-            }
-            if (!dep) {
-                MethodWellDefinedness mwd = new MethodWellDefinedness(rep, services);
-                registerContract(mwd);
-            }
-        }
-    }
-
-    /**
-     * Registers a well-definedness check for a jml statement. It does not take care of its
-     * visibility in the proof management dialog.
-     *
-     * @param swd The well-definedness check
-     */
-    public void addWdStatement(StatementWellDefinedness swd) {
-        registerWdCheck(swd);
-    }
-
-    /**
-     * Returns all registered well-definedness checks.
-     */
-    public ImmutableSet<WellDefinednessCheck> getAllWdChecks() {
-        ImmutableSet<WellDefinednessCheck> result = DefaultImmutableSet.nil();
-        for (ImmutableSet<WellDefinednessCheck> s : wdChecks.values()) {
-            result = result.union(s);
-        }
-        return result;
-    }
-
 
     // region Support SetStatement and JmlAssert
     private final Map<Statement, JmlStatementSpec> statementMap = new IdentityHashMap<>();
@@ -1873,6 +1597,11 @@ public final class SpecificationRepository {
 
     public JmlStatementSpec addStatementSpec(Statement statement, JmlStatementSpec spec) {
         return statementMap.put(statement, spec);
+    }
+
+    /// Do what ever you want.
+    /// Used by WD to modify/add represent clauses.
+    public void processJavaType(KeYJavaType kjt) {
     }
 
     /**
