+++ key.core/src/main/java/de/uka/ilkd/key/speclang/SLEnvInput.java
@@ -370,7 +370,7 @@ public final class SLEnvInput extends AbstractEnvInput {
                     specExtractor.extractMethodSpecs(constructor, staticInvPresent);
                 specRepos.addSpecs(constructorSpecs);
             }
-            specRepos.addRepresentsTermToWdChecksForModelFields(kjt);
+            specRepos.processJavaType(kjt);
         }
 
         // add initially clauses to constructor contracts
