+++ key.core/src/main/resources/META-INF/services/de.uka.ilkd.key.macros.ProofMacro
@@ -20,6 +20,5 @@ de.uka.ilkd.key.macros.AutoMacro
 de.uka.ilkd.key.macros.HeapSimplificationMacro
 de.uka.ilkd.key.macros.IntegerSimplificationMacro
 de.uka.ilkd.key.macros.OneStepProofMacro
-de.uka.ilkd.key.macros.WellDefinednessMacro
 de.uka.ilkd.key.macros.UpdateSimplificationMacro
 de.uka.ilkd.key.macros.TranscendentalFloatSMTMacro
