+++ key.core/src/main/resources/META-INF/services/de.uka.ilkd.key.proof.init.DefaultProfileResolver
@@ -4,6 +4,5 @@
 
 de.uka.ilkd.key.proof.init.JavaProfileDefaultProfileResolver
 de.uka.ilkd.key.proof.init.JavaProfileWithPermissionsDefaultProfileResolver
-de.uka.ilkd.key.wd.WdProfileResolver
 
 # de.tud.cs.se.ds.specstr.profile.StrengthAnalysisSEProfileDefaultProfileResolver
