+++ key.core/src/main/resources/META-INF/services/de.uka.ilkd.key.proof.init.loader.ProofObligationLoader
@@ -4,4 +4,3 @@ de.uka.ilkd.key.proof.init.loader.FunctionalBlockContractPOLoader
 de.uka.ilkd.key.proof.init.loader.FunctionOperationContractPOLoader
 de.uka.ilkd.key.informationflow.po.InfFlowContractPOLoader
 de.uka.ilkd.key.taclettranslation.lemma.TacletProofObligationInputLoader
-de.uka.ilkd.key.proof.init.WellDefinednessPOLoader
