+++ key.ui/src/main/java/de/uka/ilkd/key/gui/ContractSelectionPanel.java
@@ -15,8 +15,8 @@ import de.uka.ilkd.key.informationflow.impl.InformationFlowContractImpl;
 import de.uka.ilkd.key.java.Services;
 import de.uka.ilkd.key.proof.Proof;
 import de.uka.ilkd.key.speclang.*;
+import de.uka.ilkd.key.speclang.DependencyContractImpl;
 import de.uka.ilkd.key.util.LinkedHashMap;
-import de.uka.ilkd.key.wd.DependencyContractImpl;
 
 import org.key_project.util.collection.DefaultImmutableSet;
 import org.key_project.util.collection.ImmutableSet;
